requirements.txt
pyyaml
markdown
pdfkit
cvss
pygments
xlsxwriter
matplotlib